refactor: shift feedback types to src/types
