+++ src/app/modules/feedback/__tests__/feedback.service.spec.ts
@@ -8,9 +8,9 @@ import getFormFeedbackModel from 'src/app/models/form_feedback.server.model'
 
 import dbHandler from 'tests/unit/backend/helpers/jest-db'
 
+import { FeedbackResponse } from '../../../../types/form_feedback'
 import { DatabaseError } from '../../core/core.errors'
 import * as FeedbackService from '../feedback.service'
-import { FeedbackResponse } from '../feedback.types'
 
 const FormFeedback = getFormFeedbackModel(mongoose)
