+++ src/types/form_feedback.ts
@@ -2,6 +2,21 @@ import { Document, Model, QueryCursor } from 'mongoose'
 
 import { IFormSchema } from './form'
 
+export type ProcessedFeedback = {
+  index: number
+  timestamp: number
+  rating: number
+  comment: string
+  date: string
+  dateShort: string
+}
+
+export type FeedbackResponse = {
+  average?: string
+  count: number
+  feedback: ProcessedFeedback[]
+}
+
 export interface IFormFeedback {
   formId: IFormSchema['_id']
   rating: number
@@ -12,6 +27,7 @@ export interface IFormFeedbackSchema extends IFormFeedback, Document {
   created?: Date
   lastModified?: Date
 }
+
 export interface IFormFeedbackDocument extends IFormFeedbackSchema {
   created: Date
   lastModified: Date
